+++ tests/epyccel/test_builtins.py
@@ -11,8 +11,8 @@ ATOL = 1e-15
 RTOL = 2e-14
 
 
-min_int = iinfo(int).min
-max_int = iinfo(int).max
+min_int = iinfo(np.long).min
+max_int = iinfo(np.long).max
 
 min_float = finfo(float).min
 max_float = finfo(float).max
